Move extension method classes into base namespace
In following with patterns in .NET Core, move extension methods into
the "base" namespace for the type they are naturally extending so that
simply referencing the assembly will "light up" those extensions rather
than users having to also import a separate namespace.
