+++ libraries/Microsoft.Bot.Builder.Core.State/IStateManager.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Microsoft.Bot.Builder.Core.State
@@ -19,21 +18,5 @@ namespace Microsoft.Bot.Builder.Core.State
         Task Load(IEnumerable<string> keys);
 
         Task SaveChanges();
-    }
-
-    public static class StateManagerExtensions
-    {
-        public static Task<TState> Get<TState>(this IStateManager stateManager) where TState : class, new() =>
-            stateManager.Get<TState>(typeof(TState).Name);
-
-        public static Task<TState> GetOrCreate<TState>(this IStateManager stateManager) where TState : class, new() =>
-            stateManager.Get<TState>().ContinueWith(t => t.Result ?? new TState());
-
-        public static Task<TState> GetOrCreate<TState>(this IStateManager stateManager, Func<TState> stateFactory) where TState : class, new() =>
-            stateManager.Get<TState>().ContinueWith(t => t.Result ?? stateFactory());
-
-        public static void Set<TState>(this IStateManager stateManager, TState state) where TState : class =>
-            stateManager.Set(typeof(TState).Name, state);
-
     }
 }
