+++ libraries/Microsoft.Bot.Builder.Core.State/StateManagementTurnContextExtensions.cs
@@ -1,8 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
+using Microsoft.Bot.Builder.Core.State;
 
-namespace Microsoft.Bot.Builder.Core.State
+namespace Microsoft.Bot.Builder
 {
     public static class StateManagementTurnContextExtensions
     {
