+++ samples/AlarmBot/Topics/DeleteAlarmTopic.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using AlarmBot.Models;
 using AlarmBot.Responses;
-using Microsoft.Bot.Builder.Core.State;
+using Microsoft.Bot.Builder;
 using Microsoft.Bot.Schema;
 
 namespace AlarmBot.Topics
